pyproject.toml
[tool.poetry]
name = "langgraph-scratch"
version = "0.1.0"
description = ""
authors = ["amaithi-mediwave <amaithi_chirasan@mindwaveventures.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11"
langchain = "^0.1.10"
ipykernel = "^6.29.3"
langgraph = "^0.0.26"
langchainhub = "^0.1.14"
langchain-experimental = "^0.0.53"
python-dotenv = "^1.0.1"
numexpr = "^2.9.0"
grandalf = "^0.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
